Add ConfigManager
Initial commit of ConfigManager. ConfigManager is needed for effecient plugin configuration management, making sure that plugin won't access disk too often just for requesting config options. This is the home for all the plugin settings. Designed as a singleton class, it can be easily accessed across the plugin code using static getInstance() method. To get an option from the config use get() method on ConfigManager instance. Config version management and config update will be added to this class in the future. ConfigManager usage examples are included in Cerberus.php and HelpSubcommand.php