Reimplement wand giving
Wand giving is implemented through CerberusAPI class using giveWand()
method. Wand item type and enchantments can be customized
in config.yml. Wand name and lore can be customized in language files
or in config.yml. When customized in config.yml, wand-use-default-name
and wand-use-default-lore have to be set to false to make plugin
use wand-name and wand-lore options from the config.

Wand is considered a wand not because of name or lore, but because
of a special NBT tag, called isWand. It's a byte tag, but should
follow boolean logic, so 1 is true and 0 is false. isWand tag is put
into "Cerberus" compound tag in wand item's named tag. Other nbt tags
associated with Cerberus (if there will be any) should as well be put
inside "Cerberus" compound tag.

NBT tag wand consideration system allows the wand to be highly
customizable. No matter what item id, name, lore or enchantments
will a wand have, it will still remain fucntional.

- Add InventoryFullException

- Add giveWand(Player $player) method to CerberusAPI
  Throws custom InventoryFullException if wand can't be given because
  of full inventory

- Make WandSubcommand utilize CerberusAPI giveWand() method for wand
  giving

- Add wand-item, wand-enchantments, wand-use-default-name,
  wand-use-default-lore, wand-name, wand-lore options to config.yml

- Config.yml version bump

- Add wand-name, wand-lore and command.wand.inventory_full translation
  keys to eng and rus language files

- Language file version bump