Implement teleport subcommand
Teleport subcommand is used to teleport players to landclaims'
spawnpoints. Command executors can teleport themselves (if are
in-game) or other players (if they have
cerberus.command.teleport.other permission).

Landclaim is considered suitable for teleportation if:
- It exists
- It has its spawnpoint set
- Command executor owns it or, if they are not its owner, has
  cerberus.command.teleport.to.other permission
If any of these rules are not followed, command executor will
receive an error message telling the reason of unsuitability.

If landclaim name is not specified, landclaim at player's current
position will be checked for teleportation suitability and used
for teleportation. If there are multiple suitable landclaims,
command executor will receive an error message.

Details:

- Implement teleport subcommand functionality in TeleportSubcommand
  class

- Add the following permissions to plugin.yml:

  - cerberus.command.teleport.other - permission to teleport other
    players to landclaims' spawnpoints

  - cerberus.command.teleport.to.other - permission to teleport to
    other players' landclaim spawnpoints

- Add notify-player-on-teleportation option to config.yml, which is
  used to determine whether it's needed to notify player that they
  were teleported, when command executor teleports them

- Bump config.yml version to match commit number (44)

- Add translation keys and corresponding messages used by teleport
  subcommand to English and Russian language files

- Bump language files' version to match commit number (44)